Flyttet logikken for å utlede JournalpostBrevmottakere i en egen komponent
